Temporary commit.
Now I decide to implement my own collision detection again.
This time I plan to use algebraic methods, with continuous collision detection, but not
using Minkowski convolusion methods, as that is somewht messy. The planned method is
simply algebraic, using a suitably chosen in-between motion to facilitate the solving of
equations per chance.