First attempt to CCD.
Gauss map implemented now. And some basic game logic is being implemented. Next we shall
implement the computation of the Minkowski sum. After that we can test it out with some
linear motions then. And eventually we can test the performance of this implementation by
some rotating bricks.

Hope this makes the bricks game smooth and interesting.