New start.
I shall change the present approach to continuous collision detection. Instead of a
closed-form solution, an iterative algorithm is more preferable. But, in lieu of the
bilateral advancement algorithm that the author of the engine of diablo III proposes, it
might be helpful to try numerically finding the roots of the equations to solve in its
present form, instead of abandoning it completely in favor of the GJK algorithm, which in
my humble opinion is but a modified version of using Minkowski difference for collision
detection.