Only one case remains.
The remaining case is when the closest points of two polygons are two vertices. One should
then form a fixed separation axis and use the root-finding algorithm to push the deepest
points away. After this, check if any overlapping remains, if so, form new separating
functions by means of the new points and push them away again. If not, then go back to the
previous cases.