It works now.
After some debugging and adjustments the algorithm seems to be working, but it takes a lot
of time in my opinion: in some simple tests, it takes about 30~52ms to find a continuous
collision. This is kind of unacceptable. For the moment I guess it is the simulating
function which causes this tremendous consumption of time, since that simulating function
is called many times in the process. Maybe I shall optimise the way it works. On the other
hand, the situation I tested is under "high" speeds. Per chance the situation will be
better in medium-slow speeds.

If after all above optimisations the slowness problem remains, then I shall think of
optimising the way the algorithm works.