Added sequential impulses.
Now the performance is quite satisfactory. And the distance constraint (the joint) is
marvelous. Previously it was like a rope: they will be roughly constrained to certain
distances, but the "joint" is kind of soft. Now they will be rigidly fixed at prescribed
distances and the actions are interesting.

The friction is too bugy, so I remove it, until I find a way to correctly implement it.
Now it is time to make use of this engine to create a playable game.